+++ R/check_var_in_nc.R
@@ -3,7 +3,7 @@
 #' Internal function to test a custom nc file.
 #'
 #' @param bio_variables vector of names of variables to be extracted
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions.
 #'
 #' @keywords internal