+++ R/get_time_steps.R
@@ -5,7 +5,7 @@
 #' @param dataset string defining dataset to be downloaded (a list of possible
 #' values can be obtained with \code{get_available_datasets}). If set to
 #' "custom", then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included in
 #' this file.
 #'