+++ R/location_series.R
@@ -1,10 +1,14 @@
 #' Extract a time series of bioclimatic variables for one or more locations.
 #'
 #' This function extract a time series of local climate for
-#'  a set of locations
+#'  a set of locations. Note that this function does not apply any interpolation
+#'  (as opposed to \code{location_slice}). If you have a coastal location that just
+#'  falls into the water for the reconstructions, you will have to amend the coordinates
+#'  to put it more firmly on land.
 #'
-#' @param x a 2 column data.frame (with columns `longitude`, ranging
-#' -180 to 180, and `latitude`, from -90 to 90), or a vector of cell numbers.
+#' @param x a data.frame with columns `longitude`, ranging
+#' -180 to 180, and `latitude`, from -90 to 90 (and an optional `name`),
+#' or a vector of cell numbers.
 #' @param time_bp time slices in years before present (negative values represent
 #' time before present, positive values time in the future). This parameter can
 #' be a vector of times (the slices need
@@ -15,7 +19,7 @@
 #' @param bio_variables vector of names of variables to be extracted.
 #' @param dataset string defining the dataset to use. If set to "custom",
 #' then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included in
 #' this file.
 #'
@@ -39,14 +43,18 @@ location_series <-
 
     # reorder the inputs by time
     if (inherits(x, "data.frame")) {
-      locations_data <- x
+      if (!all(c("longitude","latitude") %in% names(x))){
+        stop ("x should be a dataframe with columns latitude and longitude")
+      }
+      coords <- x[,c("longitude","latitude")]
     } else if (inherits(x, "matrix"))  {
         locations_data <- as.data.frame(x) 
     } else {
       locations_data <- data.frame(cell_number = x)
     }
     
-    time_series_df <- locations_data
+    
+    time_series_df <- coords
     time_series_df$id <- seq_len(nrow(time_series_df))
     time_index <- NULL
     for (this_var in bio_variables) {
@@ -78,17 +86,20 @@ location_series <-
       }      
       # add time var if it doesn't exist yet
       if (!("time" %in% names(time_series_df))) {
-        n_time_steps <- length(time(climate_brick))
+        n_time_steps <- length(time_bp(climate_brick))
         n_locations <- nrow(time_series_df)
         time_series_df <- time_series_df[rep(
           seq_len(nrow(time_series_df)),
           n_time_steps
         ), ]
-        time_series_df$time <- rep(time(climate_brick), each = n_locations)
+        time_series_df$time <- rep(time_bp(climate_brick), each = n_locations)
       }
-      this_var_ts <- terra::extract(climate_brick, x)
-      names(this_var_ts)[-1] <- terra::time(climate_brick)
+      this_var_ts <- terra::extract(climate_brick, coords)
+      names(this_var_ts)[-1] <- time_bp(climate_brick)
       time_series_df[this_var] <- utils::stack(this_var_ts, select = -ID)$values
+    }
+    if ("name" %in% names(x)){
+      time_series_df$name <- x$name[time_series_df$id]
     }
     return(time_series_df)
   }