+++ R/location_slice.R
@@ -4,15 +4,17 @@
 #' at the appropriate times (selecting the closest time slice available for the
 #' specific date associated with each location).
 #'
-#' @param x a 2 column matrix (with columns `longitude`, ranging
-#' -180 to 180, and `latitude`, from -90 to 90), or a vector of cell numbers.
-#' @param time_bp the dates in years before present (negative
+#' @param x a data.frame with columns `longitude`, ranging
+#' -180 to 180, and `latitude`, from -90 to 90, plus optional
+#' columns `time_bp` and `name`. Alternatively, a vector of cell numbers.
+#' @param time_bp used if no `time_bp` column is present in `x`: the dates in
+#' years before present (negative
 #' values represent time before present, i.e. 1950, positive values time in the future)
 #' for each location.
 #' @param bio_variables vector of names of variables to be extracted.
 #' @param dataset string defining the dataset to use. If set to "custom",
 #' then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included in
 #' this file.
 #' @param nn_interpol boolean determining whether nearest neighbour
@@ -25,7 +27,7 @@
 
 location_slice <-
   function(x,
-           time_bp,
+           time_bp = NULL,
            bio_variables,
            dataset,
            path_to_nc = NULL,
@@ -39,21 +41,42 @@ location_slice <-
       check_var_in_nc(bio_variables, path_to_nc)
     }
 
-    # TODO CHECK THAT X HAS THE CORRECT COLUMNS AND THAT TIME_BP IS OF THE RIGHT LENGTH
-    
-    # reorder the inputs by time
-    if (inherits(x, "data.frame")) {
-      x <- x[order(time_bp), ]
-      locations_data <- x
-    } else {
-      x <- data.frame(x[order(time_bp)])
-      locations_data <- data.frame(cell_number = x[, 1])
+    # if we have a data.frame
+    if (inherits(x, "data.frame")){
+      # check that we have coordinates
+      if (!all(c("longitude","latitude") %in% names(x))){
+        stop("x must have columns latitude and longitude")
+      }
+      # check how time has been provided
+      if (("time_bp" %in% names(x)) & !is.null(time_bp)){
+        stop("times should either be given as a column of x or as values for time_bp,",
+             "not both at the same time!")
+      }
+      if (!("time_bp" %in% names(x)) & is.null(time_bp)){
+        stop("missing times: they should either be given as a column of x or as values for time_bp")
+      }
+       locations_data <- x
+       
+    } else if (inherits(x,"numeric")) {
+      locations_data <- data.frame(cell_number = x)
+      
     }
-    orig_id <- order(time_bp)
-    time_bp <- sort(time_bp)
-    locations_data$time_bp <- time_bp
+    # add time_bp if needed
+    if (!is.null(time_bp)){
+      locations_data$time_bp <- time_bp
+    }
+    #reorder input by time
+    orig_id <- order(locations_data$time_bp)
+    locations_data <- locations_data[order(locations_data$time_bp), ]
     time_indeces <- NULL
+    # store cooredinates in their own data.frame to be used for terra operations
+    if (inherits(x, "data.frame")){
+      coords <- locations_data [,c("longitude","latitude")]
+    } else {
+      coords <- locations_data [,c("cell_number")]
+    }
 
+   
     for (this_var in bio_variables) {
       # get name of file that contains this variable
       if (dataset != "custom") {
@@ -65,15 +88,16 @@ location_slice <-
         this_var_nc <- this_var
       }
       if (is.null(time_indeces)) {
+        #browser()
         times <- get_time_steps(dataset = "custom", path_to_nc = this_file)
         time_indeces <- time_bp_to_index(
-          time_bp = time_bp, time_steps = times
+          time_bp = locations_data$time_bp, time_steps = times
         )
         locations_data$time_bp_slice <- times[time_indeces]
         unique_time_indeces <- unique(time_indeces)
       }
+      
       climate_brick <- terra::rast(this_file, subds = this_var_nc)
-
       # create column to store variable
       locations_data[this_var] <- NA
       for (j in unique_time_indeces) {
@@ -81,7 +105,7 @@ location_slice <-
         this_slice_indeces <- which(time_indeces == j)
         this_climate <- terra::extract(
           x = this_slice,
-          y = x[this_slice_indeces, ]
+          y = coords[this_slice_indeces, ]
         )
         locations_data[this_slice_indeces, this_var] <- this_climate[
           ,
@@ -96,12 +120,11 @@ location_slice <-
           for (i in locations_to_move) {
             if (inherits(x, "data.frame")) {
               cell_id <-
-                terra::cellFromXY(climate_brick, as.matrix(locations_data[
+                terra::cellFromXY(climate_brick, as.matrix(coords[
                   i,
-                  1:2
                 ]))
             } else {
-              cell_id <- x[i]
+              cell_id <- coords[i]
             }
             neighbours_ids <-
               terra::adjacent(climate_brick, cell_id, 8, pairs = FALSE)