+++ R/region_series.R
@@ -16,10 +16,10 @@
 #' @param bio_variables vector of names of variables to be extracted
 #' @param dataset string defining the dataset to use. If set to "custom",
 #' then a single nc file is used from "path_to_nc"
-#' @param path_to_nc the path to the custom nc file containing the paleoclimate
+#' @param path_to_nc the path to the custom nc file containing the palaeoclimate
 #' reconstructions. All the variables of interest need to be included in
 #' this file.
-#' @param ext an extent, coded as numberic vector (length=4; 
+#' @param ext an extent, coded as numeric vector (length=4; 
 #' order= xmin, xmax, ymin, ymax) or a \code{terra::SpatExtent} object. If NULL,
 #' the full extent of the reconstruction is given.
 #' @param crop a polygon used to crop the reconstructions (e.g. the outline
@@ -106,7 +106,7 @@ region_series <-
       }
       varnames(climate_spatrasters[[this_var]]) <- this_var
       names(climate_spatrasters[[this_var]]) <- paste(this_var,
-        time(climate_spatrasters[[this_var]]),
+        time_bp(climate_spatrasters[[this_var]]),
         sep = "_"
       )
     }