+++ tests/testthat/test_download_dataset.R
@@ -10,7 +10,7 @@ test_that("download_dataset", {
     "^foo not available "
   )
   # check that only the example climate is in the data directory
-  expect_true("example_climate.nc" %in% list.files(get_data_path()))
+  expect_true("example_climate_v2.nc" %in% list.files(get_data_path()))
   # expect no error as the dataset exists
   expect_error(download_dataset("Example"), NA)
 })