+++ tests/testthat/test_get_downloaded_datasets.R
@@ -12,8 +12,9 @@ testthat::test_that("get_downloaded_datasets", {
     check_var_downloaded("npp", "Example"),
     "^npp not available"
   )
+  # in the future, use skip_if to avoid this test if we already have this variable installed
   expect_error(
-    check_var_downloaded("npp", "Krapp2021"),
+    check_var_downloaded("cloudiness_01", "Krapp2021"),
     "^variable"
   )
 })