+++ tests/testthat/test_is_region_series.R
@@ -7,13 +7,16 @@ test_that("is_region_series returns correct values", {
   expect_true(is_region_series(climate_region, strict = TRUE))
 
   # change the time for one variable
-  time(climate_region[1])<-c(1,2,3,4)
+  time(climate_region[1], tstep="years")<-c(1,2,3,4)
   # the simple test should still pass, but the strict test should not
   expect_true(is_region_series(climate_region))
   expect_false(is_region_series(climate_region, strict = TRUE))
 
   # now remove a time step (all tests should fail)
-  time(climate_region[1])<-time(climate_region[2])# reinstate correct times
+  climate_region <- region_series(
+    time_bp = list(min=-15000,max=0), c("bio01", "bio10", "bio12"),
+    "Example"
+  )
   climate_region[1] <- climate_region[1][[1:3]]
   expect_false(is_region_series(climate_region))
   expect_false(is_region_series(climate_region, strict = TRUE))